Basic-Stock-Analyzer/requirements.txt
google-adk==0.0.2
python-dotenv==1.0.0 